msm: npu: Fix format specifier warnings

+++ drivers/media/platform/msm/npu/npu_hw_access.h
@@ -62,9 +62,9 @@ uint32_t npu_bwmon_reg_read(struct npu_device *npu_dev, uint32_t off);
 void npu_bwmon_reg_write(struct npu_device *npu_dev, uint32_t off,
 	uint32_t val);
 void npu_mem_write(struct npu_device *npu_dev, void *dst, void *src,
-	uint32_t size);
+	size_t size);
 int32_t npu_mem_read(struct npu_device *npu_dev, void *src, void *dst,
-	uint32_t size);
+	size_t size);
 uint32_t npu_qfprom_reg_read(struct npu_device *npu_dev, uint32_t off);
 
 int npu_mem_map(struct npu_client *client, int buf_hdl, uint32_t size,
